Do not sanitize missing request properties
This commit fixes a potential bug caused by the fact that sanitizers
would run, regardless of whether the request value for a given sanitizer
exists.

In most cases, this bug will probably not be noticed, but I can imagine
this failing if one were to add a sanitizer that prefixes every value
with a string (or something like that), causing the request values to be
polluted with values that are not expected.

This commit fixes #12