Use Laravel's prepareForValidation instead of hooking into $this->all()
This commit changes the behavior to be more compliant with the way
Laravel's internal validation works.

After a FormRequest instance has resolved in the application container,
the validator will be automatically called. Before the validation runs,
"prepareForValidation()" is triggered which has the purpose of alterting
request data before it enters validation (which, is exactly what this
package does!)

This commit also borrows a lot of test scaffolding from Laravel's
internal testing, because it is not trivial to create a FormRequest
object without initializing the entire application.